scripts/livekit_agent/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-cartesia
livekit-plugins-silero
python-dotenv
